+++ R/bioconductor.R
@@ -31,21 +31,41 @@ renv_bioconductor_init_biocinstaller <- function() {
 
 }
 
-renv_bioconductor_version <- function() {
+renv_bioconductor_version <- function(project) {
 
-  if (renv_package_available("BiocManager")) {
-    BiocManager <- renv_namespace_load("BiocManager")
-    format(BiocManager$version())
-  } else if (renv_package_available("BiocInstaller")) {
-    BiocInstaller <- renv_namespace_load("BiocInstaller")
-    format(BiocInstaller$biocVersion())
-  } else if (renv_package_available("BiocVersion")) {
-    format(packageVersion("BiocVersion")[1, 1:2])
-  }
+  # check and see if we have an override via option
+  version <- getOption("renv.bioconductor.version")
+  if (!is.null(version))
+    return(version)
+
+  # check and see if the project has been configured to use a specific
+  # Bioconductor release
+  version <- settings$bioconductor.version(project = project)
+  if (length(version))
+    return(version)
+
+  # otherwise, infer the Bioconductor version from installed packages
+  case(
+
+    renv_package_available("BiocVersion") ~ {
+      format(packageVersion("BiocVersion")[1, 1:2])
+    },
+
+    renv_package_available("BiocManager") ~ {
+      BiocManager <- renv_namespace_load("BiocManager")
+      format(BiocManager$version())
+    },
+
+    renv_package_available("BiocVersion") ~ {
+      BiocInstaller <- renv_namespace_load("BiocInstaller")
+      format(BiocInstaller$biocVersion())
+    }
+
+  )
 
 }
 
-renv_bioconductor_repos <- function(version = NULL) {
+renv_bioconductor_repos <- function(project, version = NULL) {
 
   # allow bioconductor repos override
   repos <- getOption("renv.bioconductor.repos")
@@ -53,7 +73,7 @@ renv_bioconductor_repos <- function(version = NULL) {
     return(repos)
 
   # read Bioconductor version (normally set during restore)
-  version <- version %||% getOption("renv.bioconductor.version")
+  version <- version %||% renv_bioconductor_version(project = project)
 
   # try both BiocManager, BiocInstaller to get Bioconductor repositories
   getters <- list(